refactor(codegen): split to codegen
